Przerobienie projektu bazy danych z brancha "Database" na "pakietową" formę projektu aplikacji.
