--- user_database.py
@@ -1,14 +1,0 @@
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from app import app
-
-db = SQLAlchemy(app)
-
-
-class User(db.Model):
-    __tablename__ = 'user_data'
-    id = db.Column(db.Integer, primary_key=True)
-    town = db.Column(db.String)
-
-    def __init__(self):
-        self.town = 'Lazy Town'
